+++ STACpopulator/exceptions.py
@@ -1,5 +1,10 @@
 class STACPopulatorError(Exception):
+    """STAC populator Error."""
+
     pass
 
+
 class ExtensionLoadError(STACPopulatorError):
+    """Extension Loading Error."""
+
     pass