+++ STACpopulator/extensions/base.py
@@ -1,5 +1,5 @@
 """
-# Base classes for STAC extensions
+Base classes for STAC extensions.
 
 What we have:
   - `Loader`, which returns attributes.
@@ -22,35 +22,44 @@ How-to:
 
 
 """
+
 from __future__ import annotations
 
-from abc import ABC, abstractmethod
-from datetime import datetime as Datetime
 import json
-import jsonschema
 import logging
-from typing import Any, Dict, Generic,    TypeVar, Union, cast, Optional
-from pydantic import (BaseModel, create_model, Field, FilePath, field_validator, model_validator, HttpUrl, ConfigDict,
-                      PrivateAttr)
+import tempfile
+import types
+from abc import abstractmethod
+from datetime import datetime as Datetime
+from pathlib import Path
+from typing import Any, Generic, Optional, Type, TypeVar, Union, cast
+
+import jsonschema
 import pystac
+from pydantic import (
+    BaseModel,
+    ConfigDict,
+    Field,
+    FilePath,
+    PrivateAttr,
+    create_model,
+    model_validator,
+)
+from pystac import STACValidationError
 from pystac.extensions import item_assets
 from pystac.extensions.base import (
     ExtensionManagementMixin,
     PropertiesExtension,
-    SummariesExtension,
+    S,  # generic pystac.STACObject
 )
-from pystac import STACValidationError
-from pystac.extensions.base import S  # generic pystac.STACObject
-from STACpopulator.models import AnyGeometry, GeoJSONPolygon
+
+from STACpopulator.models import AnyGeometry
 from STACpopulator.stac_utils import ServiceType
-import types
-import uuid
-from pathlib import Path
-import tempfile
 
 T = TypeVar("T", pystac.Collection, pystac.Item, pystac.Asset, item_assets.AssetDefinition)
 
 LOGGER = logging.getLogger(__name__)
+
 
 class Helper:
     """Class to be subclassed by extension helpers."""
@@ -70,6 +79,7 @@ class ExtensionHelper(BaseModel, Helper):
     _schema_exclude : list[str]
         Properties not meant to be validated by json schema, but still included in the data model.
     """
+
     _prefix: str = PrivateAttr()
     _schema_uri: FilePath = PrivateAttr(None)
     _schema_exclude: list[str] = PrivateAttr([])
@@ -77,7 +87,7 @@ class ExtensionHelper(BaseModel, Helper):
     model_config = ConfigDict(populate_by_name=True, extra="ignore", ser_json_inf_nan="strings")
 
     @classmethod
-    def __init_subclass__(cls, **kwargs):
+    def __init_subclass__(cls, **kwargs) -> None:
         """Automatically set an alias generator from the `_prefix`."""
         prefix = cls._prefix.default
 
@@ -86,10 +96,10 @@ class ExtensionHelper(BaseModel, Helper):
 
     @model_validator(mode="before")
     @classmethod
-    def validate_jsonschema(cls, data):
+    def validate_jsonschema(cls, data: dict) -> dict:
         """Validate the data model against the json schema, if given."""
         # Load schema
-        uri =  cls._schema_uri.default
+        uri = cls._schema_uri.default
         if uri is not None:
             with open(uri) as f:
                 schema = json.load(f)
@@ -97,15 +107,17 @@ class ExtensionHelper(BaseModel, Helper):
             validator_cls.check_schema(schema)
             validator = validator_cls(schema)
 
-            attrs = {f"{cls._prefix.default}:{k}": v for (k,v) in data.items() if k not in cls._schema_exclude.default}
+            attrs = {f"{cls._prefix.default}:{k}": v for (k, v) in data.items() if k not in cls._schema_exclude.default}
             errors = list(validator.iter_errors(attrs))
             if errors:
                 raise ValueError(errors)
 
         return data
 
-    def apply(self, item, add_if_missing=True):
-        """Add extension for the properties of the dataset to the STAC item.
+    def apply(self, item: pystac.Item, add_if_missing: bool = True) -> pystac.Item:
+        """
+        Add extension for the properties of the dataset to the STAC item.
+
         The extension class is created dynamically from the properties.
         """
         schema_uri = self.write_stac_schema() if self._schema_uri else None
@@ -120,13 +132,14 @@ class ExtensionHelper(BaseModel, Helper):
 
     def to_stac_schema(self) -> dict:
         """Return the STAC schema for the extension."""
-        return {'type': 'object',
-                'required': ['type', 'properties'],
-                'properties': {'type': {'const': 'Feature'},
-                               'properties': {'$ref': str(self._schema_uri)}
-                               }
-                }
+        return {
+            "type": "object",
+            "required": ["type", "properties"],
+            "properties": {"type": {"const": "Feature"}, "properties": {"$ref": str(self._schema_uri)}},
+        }
+
     def write_stac_schema(self) -> str:
+        """Write STAC schema to a temporary file and return that file path."""
         path = Path(tempfile.mkdtemp()) / f"{self._prefix}-schema.json"
         with open(path, "w") as fh:
             json.dump(self.to_stac_schema(), fh)
@@ -149,6 +162,7 @@ class BaseSTAC(BaseModel):
     id : str
         The unique identifier of the dataset. If not set, it should be generated by a subclass' `create_uid`.
     """
+
     # STAC item properties
     geometry: AnyGeometry | None
     bbox: list[float]
@@ -159,23 +173,22 @@ class BaseSTAC(BaseModel):
 
     model_config = ConfigDict(populate_by_name=True, extra="ignore", arbitrary_types_allowed=True)
 
-
     # Helpers are automatically detected by being Helper subclasses
     _helpers: list[str] = PrivateAttr([])
 
     @model_validator(mode="after")
-    def set_id(self):
-        """Set the ID of the dataset if None"""
+    def set_id(self) -> "BaseSTAC":
+        """Set the ID of the dataset if None."""
         if self.id is None:
             self.id = self.create_uid()
         return self
 
     @abstractmethod
-    def create_uid(self):
+    def create_uid(self) -> str:
         """Return a unique identifier for the dataset."""
 
     @model_validator(mode="after")
-    def find_helpers(self):
+    def find_helpers(self) -> "BaseSTAC":
         """Populate the list of extensions."""
         for key, field in self.model_fields.items():
             if isinstance(field.annotation, type) and issubclass(field.annotation, Helper):
@@ -206,28 +219,29 @@ class BaseSTAC(BaseModel):
         return json.loads(json.dumps(item.to_dict()))
 
 
-def metacls_extension(name, schema_uri):
+def metacls_extension(name: str, schema_uri: str) -> "Type[MetaExtension]":
     """Create an extension class dynamically from the properties."""
     cls_name = f"{name.upper()}Extension"
 
-    bases = (MetaExtension,
-             Generic[T],
-             PropertiesExtension,
-             ExtensionManagementMixin[Union[pystac.Asset, pystac.Item, pystac.Collection]]
-             )
+    bases = (
+        MetaExtension,
+        Generic[T],
+        PropertiesExtension,
+        ExtensionManagementMixin[Union[pystac.Asset, pystac.Item, pystac.Collection]],
+    )
 
     attrs = {"name": name, "schema_uri": schema_uri}
     return types.new_class(name=cls_name, bases=bases, kwds=None, exec_body=lambda ns: ns.update(attrs))
 
 
 class MetaExtension:
+    """Extension metaclass."""
+
     name: str
     schema_uri: str
 
     def apply(self, properties: dict[str, Any]) -> None:
-        """Applies Extension properties to the extended
-        :class:`~pystac.Item` or :class:`~pystac.Asset`.
-        """
+        """Apply Extension properties to the extended :class:`~pystac.Item` or :class:`~pystac.Asset`."""
         for prop, val in properties.items():
             self._set_property(prop, val)
 
@@ -237,20 +251,20 @@ class MetaExtension:
         return cls.schema_uri
 
     @classmethod
-    def has_extension(cls, obj: S):
+    def has_extension(cls, obj: S) -> bool:
+        """Return True iff the object has an extension for that matches this class' schema URI."""
         ext_uri = cls.get_schema_uri()
         return obj.stac_extensions is not None and any(uri == ext_uri for uri in obj.stac_extensions)
 
     @classmethod
-    def ext(cls, obj: T, add_if_missing: bool = False) -> "Extension[T]":
-        """Extends the given STAC Object with properties from the
-        :stac-ext:`Extension`.
+    def ext(cls, obj: T, add_if_missing: bool = False) -> "Extension[T]":  # noqa: F821
+        """Extend the given STAC Object with properties from the :stac-ext:`Extension`.
 
         This extension can be applied to instances of :class:`~pystac.Item` or
         :class:`~pystac.Asset`.
 
-        Raises:
-
+        Raises
+        ------
             pystac.ExtensionTypeError : If an invalid object type is passed.
         """
         cls_map = {pystac.Item: MetaItemExtension}
@@ -264,7 +278,7 @@ class MetaExtension:
             raise pystac.ExtensionTypeError(cls._ext_error_message(obj))
 
 
-def extend_type(stac, cls, ext):
+def extend_type(stac: T, cls: MetaItemExtension, ext: MetaItemExtension[T]) -> Type:
     """Create an extension subclass for different STAC objects.
 
     Note: This is super confusing... we should come up with some better nomenclature.
@@ -278,19 +292,21 @@ def extend_type(stac, cls, ext):
     ext: MetaExtension[T]
       The meta extension class.
     """
-    cls_name = f"{stac.__name__ }{ext.__name__}"
+    cls_name = f"{stac.__name__}{ext.__name__}"
     return types.new_class(cls_name, (cls, ext), {}, lambda ns: ns)
 
 
 class MetaItemExtension:
-    """A concrete implementation of :class:`Extension` on an :class:`~pystac.Item`
-    that extends the properties of the Item to include properties defined in the
+    """A concrete implementation of :class:`Extension` on an :class:`~pystac.Item`.
+
+    Extends the properties of the Item to include properties defined in the
     :stac-ext:`Extension`.
 
     This class should generally not be instantiated directly. Instead, call
     :meth:`Extension.ext` on an :class:`~pystac.Item` to extend it.
     """
-    def __init__(self, item: pystac.Item):
+
+    def __init__(self, item: pystac.Item) -> None:
         self.item = item
         self.properties = item.properties
 
@@ -304,7 +320,8 @@ class MetaItemExtension:
             service_type: If set, filter the assets such that only those with a
                 matching :class:`~STACpopulator.stac_utils.ServiceType` are returned.
 
-        Returns:
+        Returns
+        -------
             Dict[str, Asset]: A dictionary of assets that match ``service_type``
                 if set or else all of this item's assets were service types are defined.
         """
@@ -316,10 +333,12 @@ class MetaItemExtension:
         }
 
     def __repr__(self) -> str:
+        """Return repr."""
         return f"<{self.__class__.__name__} Item id={self.item.id}>"
 
 
 # TODO: Add the other STAC item meta extensions
+
 
 def schema_properties(schema: dict) -> list[str]:
     """Return the list of properties described by schema."""
@@ -330,10 +349,17 @@ def schema_properties(schema: dict) -> list[str]:
     return out
 
 
-def model_from_schema(model_name, schema: dict):
+def model_from_schema(model_name: str, schema: dict) -> BaseModel:
     """Create pydantic BaseModel from JSON schema."""
-    type_map = {"string": str, "number": float, "integer": int, "boolean": bool, "array": list, "object": dict,
-                None: Any}
+    type_map = {
+        "string": str,
+        "number": float,
+        "integer": int,
+        "boolean": bool,
+        "array": list,
+        "object": dict,
+        None: Any,
+    }
 
     fields = {}
     for key, val in schema["properties"].items():
@@ -342,4 +368,3 @@ def model_from_schema(model_name, schema: dict):
         default = ... if key in schema["required"] else None
         fields[name] = (typ, Field(default, alias=key))
     return create_model(model_name, **fields)
-