+++ STACpopulator/extensions/cordex6.py
@@ -1,8 +1,9 @@
 from __future__ import annotations
 
-from pathlib import Path
-from pydantic import BaseModel, Field, FilePath, model_validator
 from datetime import datetime
+from pathlib import Path
+
+from pydantic import FilePath, model_validator
 
 from STACpopulator.extensions.base import ExtensionHelper
 from STACpopulator.extensions.thredds import THREDDSCatalogDataModel
@@ -11,6 +12,8 @@ from STACpopulator.extensions.xscen import Xscen
 
 # This is generated using datamodel-codegen + manual edits
 class CordexCmip6(ExtensionHelper):
+    """Helper for CORDEX CMIP6 data."""
+
     # Fields from schema
     activity_id: str
     contact: str
@@ -46,10 +49,10 @@ class CordexCmip6(ExtensionHelper):
     _schema_uri: FilePath = Path(__file__).parent / "schemas" / "cordex6" / "cmip6-cordex-global-attrs-schema.json"
 
 
-
 # Customize the THREDDSCatalogDataModel
 class Cordex6DataModel(THREDDSCatalogDataModel):
     """Data model for CORDEX-CMIP6 NetCDF datasets."""
+
     cordex6: CordexCmip6
 
     def create_uid(self) -> str:
@@ -72,19 +75,21 @@ class Cordex6DataModel(THREDDSCatalogDataModel):
 
     @model_validator(mode="before")
     @classmethod
-    def properties_helper(cls, data):
+    def properties_helper(cls, data: dict) -> dict:
         """Instantiate the properties helper."""
-        data["cordex6"] = data['data']['attributes']
+        data["cordex6"] = data["data"]["attributes"]
         return data
 
 
 # Customize the THREDDSCatalogDataModel
 class Cordex6DataModelNcML(Cordex6DataModel):
     """Data model for CORDEX-CMIP6 NcML aggregations."""
+
     xscen: Xscen
 
     @model_validator(mode="before")
     @classmethod
-    def xscen_helper(cls, data):
-        data['xscen'] = data['data']['attributes']
+    def xscen_helper(cls, data: dict) -> dict:
+        """Instantiate the XSCEN helper."""
+        data["xscen"] = data["data"]["attributes"]
         return data