+++ STACpopulator/extensions/xscen.py
@@ -1,10 +1,17 @@
 from __future__ import annotations
 
 from typing import Literal
+
 from STACpopulator.extensions.base import ExtensionHelper
 
 
 class Xscen(ExtensionHelper):
+    """
+    XSCEN extension helper class.
+
+    See: https://github.com/Ouranosinc/xscen
+    """
+
     type: Literal["forecast", "station-obs", "gridded-obs", "reconstruction", "simulation"]
     processing_level: Literal["raw", "extracted", "regridded", "biasadjusted"]
     license_type: Literal["permissive", "permissive non-commercial"]