+++ STACpopulator/implementations/CORDEXCMIP6_Ouranos/add_CORDEX6.py
@@ -1,31 +1,34 @@
+import argparse
 import logging
+from typing import Any
+
 from requests.sessions import Session
 
-from STACpopulator.log import add_logging_options
-from STACpopulator.request_utils import add_request_options, apply_request_options
-from STACpopulator.input import ErrorLoader, THREDDSLoader
-
-import argparse
-from typing import Any
-from STACpopulator.populator_base import STACpopulatorBase
 from STACpopulator.extensions.cordex6 import Cordex6DataModel
+from STACpopulator.input import ErrorLoader, THREDDSLoader
+from STACpopulator.log import add_logging_options
+from STACpopulator.populator_base import STACpopulatorBase
+from STACpopulator.request_utils import add_request_options, apply_request_options
 
 LOGGER = logging.getLogger(__name__)
 
 
 class CORDEX_STAC_Populator(STACpopulatorBase):
+    """Populator that creates STAC objects representing CORDEX data from a THREDDS catalog."""
+
     data_model = Cordex6DataModel
     item_geometry_model = None  # Unnecessary, but kept for consistency
 
     def create_stac_item(self, item_name: str, item_data: dict[str, Any]) -> dict[str, Any]:
+        """Return a STAC item."""
         dm = self.data_model.from_data(item_data)
         return dm.stac_item()
 
 
-
 # TODO: This probably doesn't need to be copied for every implementation, right ?
 def add_parser_args(parser: argparse.ArgumentParser) -> None:
-    parser.description="CMIP6-CORDEX STAC populator from a THREDDS catalog or NCML XML."
+    """Add additional CLI arguments to the argument parser."""
+    parser.description = "CMIP6-CORDEX STAC populator from a THREDDS catalog or NCML XML."
     parser.add_argument("stac_host", help="STAC API URL")
     parser.add_argument("href", help="URL to a THREDDS catalog or a NCML XML with CMIP6 metadata.")
     parser.add_argument("--update", action="store_true", help="Update collection and its items")
@@ -48,6 +51,7 @@ def add_parser_args(parser: argparse.ArgumentParser) -> None:
 
 
 def runner(ns: argparse.Namespace) -> int:
+    """Run the populator."""
     LOGGER.info(f"Arguments to call: {vars(ns)}")
 
     with Session() as session: