+++ tests/test_base.py
@@ -2,9 +2,9 @@
 
 from __future__ import annotations
 
-from typing import Literal
-from STACpopulator.extensions.base import ExtensionHelper, BaseSTAC
 import datetime as dt
+
+from STACpopulator.extensions.base import BaseSTAC, ExtensionHelper
 
 
 class ExTest(ExtensionHelper):
@@ -16,6 +16,7 @@ class ExTest(ExtensionHelper):
 # Customize the THREDDSCatalogDataModel
 class TestDataModel(BaseSTAC):
     """Data model for CORDEX-CMIP6 NetCDF datasets."""
+
     ex: ExTest
 
     def create_uid(self) -> str:
@@ -24,19 +25,12 @@ class TestDataModel(BaseSTAC):
 
 
 def test_extension():
-    geom =  {
-        "type": "Polygon",
-        "coordinates": [[[0, 0], [0, 1], [1, 1], [1, 0], [0, 0]]]
-    }
-    model = TestDataModel(id="test",
-                  geometry=geom,
-                  bbox=[0,0,1,1],
-                  datetime=dt.datetime(2021,1,1),
-                  ex= {"f": 1.0, "n": float("nan")})
+    geom = {"type": "Polygon", "coordinates": [[[0, 0], [0, 1], [1, 1], [1, 0], [0, 0]]]}
+    model = TestDataModel(
+        id="test", geometry=geom, bbox=[0, 0, 1, 1], datetime=dt.datetime(2021, 1, 1), ex={"f": 1.0, "n": float("nan")}
+    )
 
     assert set(model._helpers) == {"ex"}
     item = model.stac_item()
     assert item["properties"]["ex:f"] == 1.0
     assert item["properties"]["ex:n"] == "NaN"
-
-