+++ tests/test_standalone_stac_item.py
@@ -1,15 +1,15 @@
 import json
-import pystac
-import pytest
-import requests
 import os
 import tempfile
 from urllib.parse import quote
 
+import pystac
+import pytest
+import requests
 import xncml
 
 from STACpopulator.extensions.cmip6 import CMIP6Helper
-from STACpopulator.extensions.thredds import THREDDSHelper, THREDDSExtension
+from STACpopulator.extensions.thredds import THREDDSExtension, THREDDSHelper
 from STACpopulator.implementations.CMIP6_UofT.add_CMIP6 import CMIP6populator
 from STACpopulator.input import THREDDSLoader
 from STACpopulator.models import GeoJSONPolygon
@@ -65,7 +65,7 @@ class MockedNoSTACUpload(CMIP6populator):
             "keywords": ["test"],
             "license": "MIT",
             "spatialextent": [-180, -90, 180, 90],
-            "temporalextent": ['1850-01-01', None]
+            "temporalextent": ["1850-01-01", None],
         }
 
     def validate_host(self, stac_host: str) -> str: